% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edashop-package.R
\docType{data}
\name{auctions_phy}
\alias{auctions_phy}
\title{Auctions in a Distressed Real Estate Market in Italy: Physical Features of Property.}
\format{
A data frame with 125 rows and 15 variables:
\describe{
\item{id}{Unique identifier of the property auctioned}
\item{type_class}{Type of auctioned property; factor with 7 levels (Factory, Residence, Retail, Mixed, Build-on Land, Office, Agricultural Building)}
\item{gross_building_area}{Gross Building Area (m2)}
\item{quality}{An ordinal factor that describes the quality of the property, including design work and materials employed (Poor, Adequate, Fair, Good, Excellent)}
\item{state_maintenance}{An ordinal factor that describes the state of maintenance of the property, if it is new, recently renovated or damaged over the years (Poor, Adequate, Fair, Good, Excellent)}
}
}
\source{
Canesi, Rubina, and Giuliano Marella. "Data from RE distressed market: Properties auctions in Italy." Data in Brief 18 (2018): 319-324. \doi{10.1016/j.dib.2018.03.009}
}
\usage{
data(auctions_phy)
}
\description{
Data describing Real Estate (RE) distressed market, with a focus on property foreclosures in North-East Italy between 2000 and 2016. The data
were collected by Canesi and Marella by means of a survey sent to financial institutions, courts of law and different associations of public
notaries. The aim of the survey was to record real estate auctions, and their technical and socio-economic features. The data provide information
about housing market trends and performance as well as demographic features.
}
\details{
This table contains the physical features of the properties auctioned.
}
\examples{
 data(auctions_phy)
 quality <- auctions_phy$quality
}
\keyword{datasets}
