% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edashop-package.R
\docType{data}
\name{cntr_sp_basico}
\alias{cntr_sp_basico}
\title{Basico - Census tracts}
\format{
A simple features data frame with 66,096 rows and 30 variables:
\describe{
\item{code_tract}{Census tract code}
\item{situacao}{Census tract status code}
\item{b_V001}{Permanent private households or heads of permanent private households}
\item{b_V002}{Residents in permanent private households or population living in permanent private households}
\item{b_V003}{Average number of residents in permanent private households}
\item{b_V004}{Variance in the number of residents in permanent private households}
\item{b_V005}{Average nominal monthly income of heads of household}
}
}
\source{
CENSOBR R, retrieved using https://ipeagit.github.io/censobr/index.html
}
\usage{
data(cntr_sp_basico)
}
\description{
Variables from the Head of household Questionnaire of the 2010 Brazilian Demographic Census, applied to all household units.
}
\examples{
 data(cntr_sp_ipvs)
 situacao <- cntr_sp_ipvs$situacao
}
\keyword{datasets}
