% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edashop-package.R
\docType{data}
\name{cntr_sp_geo}
\alias{cntr_sp_geo}
\title{Census tracts geometry}
\format{
A simple features data frame with 66,096 rows and 5 variables:
\describe{
\item{COD_SETOR}{Census tract code}
\item{code_muni}{Municipality code}
\item{name_muni}{Municipality name}
\item{code_state}{State code}
\item{geometry}{Census tracts geometry}
}
}
\source{
CENSOBR R, https://ipeagit.github.io/censobr/index.html
}
\usage{
data(cntr_sp_geo)
}
\description{
Simple features object with census tracts of São Paulo State.
}
\examples{
 data(cntr_sp_geo)
 codes <- cntr_sp_geo$COD_SETOR
}
\keyword{datasets}
