% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edashop-package.R
\docType{data}
\name{cntr_sp_head}
\alias{cntr_sp_head}
\title{Head of household - Census tracts}
\format{
A simple features data frame with 66,096 rows and 11 variables:
\describe{
\item{code_tract}{Census tract code}
\item{h_V001}{Heads of household with a nominal monthly income of up to 1/2 the minimum wage}
\item{h_V002}{Heads of household with nominal monthly income of more than 1/2 to 1 minimum wage}
\item{h_V003}{Heads of household with nominal monthly income of more than 1 to 2 minimum wages}
\item{h_V004}{Heads of household with nominal monthly income of more than 2 to 3 minimum wages}
\item{h_V005}{Heads of household with a nominal monthly income of more than 3 to 5 minimum wages}
\item{h_V006}{Heads of household with a nominal monthly income of more than 5 to 10 minimum wages}
\item{h_V007}{Heads of household with a nominal monthly income of more than 15 to 20 minimum wages}
\item{h_V008}{Heads of household with a nominal monthly income of more than 15 to 20 minimum wages}
\item{h_V009}{Heads of household with a nominal monthly income of more than 20 minimum wages}
\item{h_V010}{Heads of household with no nominal monthly income}
}
}
\source{
CENSOBR R, https://ipeagit.github.io/censobr/index.html
}
\usage{
data(cntr_sp_head)
}
\description{
Variables from the Head of household Questionnaire of the 2010 Brazilian Demographic Census, applied to all household units.
}
\examples{
 data(cntr_sp_head)
 v001 <- cntr_sp_head$h_V001
}
\keyword{datasets}
