% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edashop-package.R
\name{eda_session_2_br}
\alias{eda_session_2_br}
\title{EDA Workshop Session 2 Template (Brazilian Edition).}
\format{
An R Markdown file
}
\usage{
eda_session_2_br()
}
\description{
Template for creating a computational notebook for Session 2 of the Brazilian
edition of the workshop.
}
\keyword{datasets}
