% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edashop-package.R
\name{eda_session_4_it}
\alias{eda_session_4_it}
\title{EDA Workshop Session 4 Template (Italian Edition).}
\format{
An R Markdown file
}
\usage{
eda_session_4_it()
}
\description{
Template for creating a computational notebook for Session 4 of the Italian
edition of the workshop.
}
\keyword{datasets}
