% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edashop-package.R
\name{eda_session_5}
\alias{eda_session_5}
\title{EDA Workshop Session 5 Template (All Editions).}
\format{
An R Markdown file
}
\usage{
eda_session_5()
}
\description{
Template for creating a computational notebook for Session 5 of the workshop.
}
\keyword{datasets}
