% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edashop-package.R
\docType{data}
\name{i40_index_rank}
\alias{i40_index_rank}
\title{Promethee Rankings of Industry Readiness Index 4.0.}
\format{
A data frame with 309 rows and 4 variables:
\describe{
\item{NUTS_ID}{NUTS-2 identifier of region}
\item{gdp_rank}{GDP rank of the region}
\item{promethee_rank}{Industry Readiness Index rank of the region}
\item{regional_innovation_index}{2019 Regional Innovation Index rank of the region}
}
}
\source{
Honti, Gergely, Tímea Czvetko, and Janos Abonyi. "Data describing the regional Industry 4.0 readiness index." Data in Brief 33 (2020): 106464. https://doi.org/10.1016/j.dib.2020.106464
}
\usage{
data(i40_index_rank)
}
\description{
A table with rankings of the Readiness Index 4.0 obtained using the Promethee method. For details on the inputs see \code{\link{i40_index_db}}
}
\examples{
 data(i40_index_rank)
 NUTS_2 <- i40_index_rank$NUTS_ID
}
\keyword{datasets}
