% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edashop-package.R
\docType{data}
\name{italy_auctions}
\alias{italy_auctions}
\title{Auctions in a Distressed Real Estate Market in Italy.}
\format{
A data frame with 125 rows and 15 variables:
\describe{
\item{type_class}{Type of auctioned property; factor with 7 levels (Factory, Residence, Retail, Mixed, Build-on Land, Office, Agricultural Building)}
\item{gross_building_area}{Gross Building Area (m2)}
\item{quality}{An ordinal factor that describes the quality of the property, including design work and materials employed (Poor, Adequate, Fair, Good, Excellent)}
\item{state_maintenance}{An ordinal factor that describes the state of maintenance of the property, if it is new, recently renovated or damaged over the years (Poor, Adequate, Fair, Good, Excellent)}
\item{occupancy}{The state of ccupancy of the auctioned property as an indicator of whether the property generates income; factor with 3 levels (Unoccupied, Tenant, Owner)}
\item{location}{Physical location based on distance from the city center of property, ordinal factor (Center, Semi-center, Suburban)}
\item{income}{Average per capita income for the city where the auction takes place (euros/year)}
\item{delta_ntn}{Variation in the number of local real estate transactions over the observed period, an index that describes the activity of the local real estate market (percentage)}
\item{re_activity_index}{Real Estate Activity Index (REAI) is an index that describes the activity and the performance of the local real estate market. This index is calculated as the ratio between delta_ntn and the total stock of real estate units in the area where the property is auctioned}
\item{population}{Population in the city where the property is auctioned}
\item{days_on_market}{Number of days the property stayed on the auctioned market before being sold; the difference between the date of sale (closing date) and the first auction starting date (opening bid)}
\item{number_auctions}{The number of bidding proceedings before the selling}
\item{discount}{Percentage variation between the first listing value (which equals the market value assessed by the appraisal) and the selling price }
\item{premium}{Premium paid by the winner bidder to win the auction calculated as the percentage variation between the last listing value and the final selling price}
\item{date}{Date when the property was assessed by the appraisal, in Year-Month-Day format}
}
}
\source{
Canesi, Rubina, and Giuliano Marella. "Data from RE distressed market: Properties auctions in Italy." Data in Brief 18 (2018): 319-324. \url{https://doi.org/10.1016/j.dib.2018.03.009}
}
\usage{
data(italy_auctions)
}
\description{
Data describing Real Estate (RE) distressed market, with a focus on property foreclosures in North-East Italy between 2000 and 2016. The data
were collected by Canesi and Marella by means of a survey sent to financial institutions, courts of law and different associations of public
notaries. The aim of the survey was to record real estate auctions, and their technical and socio-economic features. The data provide information
about housing market trends and performance as well as demographic features.
}
\examples{
 data(italy_auctions)
 gradient <- italy_auctions$grad_existing
}
\keyword{datasets}
