% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edashop-package.R
\docType{data}
\name{italy_nuts2}
\alias{italy_nuts2}
\title{NUTS-2 regions in Italy.}
\format{
A data frame with 21 rows and 9 variables:
\describe{
\item{NUTS_ID}{NUTS-2 identifier of region}
\item{LEVL_CODE}{Level of NUTS (2)}
\item{CNTR_CODE}{Code of country (IT)}
\item{NUTS_NAME}{Name of region}
\item{NAME_LATN}{Name of region in Latin script}
\item{FID}{NUTS-2 identifier of region}
\item{geometry}{Geometry of regions}
}
}
\source{
Eurostat, retrieved with \code{\link[eurostat]{get_eurostat_geospatial}}
}
\usage{
data(italy_nuts2)
}
\description{
Simple features object with geometry of NUTS-2 regions in  Italy (unprojected).
}
\examples{
 data(italy_nuts2)
 NAME_LATN <- italy_nuts2$NAME_LATN
}
\keyword{datasets}
