% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edashop-package.R
\docType{data}
\name{nuts2}
\alias{nuts2}
\title{NUTS-2 regions.}
\format{
A simple features data frame with 320 rows and 7 variables:
\describe{
\item{NUTS_ID}{NUTS-2 identifier of region}
\item{LEVL_CODE}{Level of NUTS (2)}
\item{CNTR_CODE}{Code of country (IT)}
\item{NUTS_NAME}{Name of region}
\item{NAME_LATN}{Name of region in Latin script}
\item{FID}{NUTS-2 identifier of region}
\item{geometry}{Geometry of regions}
}
}
\source{
Eurostat, retrieved with \code{\link[eurostat]{get_eurostat_geospatial}}
}
\usage{
data(nuts2)
}
\description{
Simple features object with geometry of NUTS-2 regions (unprojected).
}
\examples{
 data(nuts2)
 NAME_LATN <- nuts2$NAME_LATN
}
\keyword{datasets}
