% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edashop-package.R
\docType{data}
\name{phd_italy_regions}
\alias{phd_italy_regions}
\title{Regional summary of survey on Ph.D. entrepreneurship in Italy.}
\format{
A data frame with 19 rows and 7 variables:
\describe{
\item{geo}{Geographical location (Northern Italy, Central Italy, Southern Italy)}
\item{nome_regione}{Name of region}
\item{NUTS_ID}{NUTS-2 identifier of region}
\item{phd_students}{Number of respondents to survey}
\item{active_spinoff}{Number of active spinoffs founded by respondents to survey}
\item{employees}{Number of people employed in spinoffs founded by respondents to survey}
\item{spinoff0506}{Spinoffs in 2005-06 in the region where the startup was established}
\item{population05}{Total population of region in 2005}
}
}
\source{
Muscio, Alessandro, and Laura Ramaciotti. "Dataset from a qualitative survey on Ph. D. entrepreneurship in Italy." Data in Brief 18 (2018): 1272-1276. \doi{10.1016/j.dib.2018.03.116}
}
\usage{
data(phd_italy_regions)
}
\description{
Regional summary of selected variables drawn from Muscio and Ramaciotti (2018). The original data are from a survey on Ph.D. (Doctor of Philosophy)
entrepreneurship in Italy. For further detail on this data set, see \code{\link{phd_italy}}
}
\examples{
 data(phd_italy_regions)
 geo <- phd_italy_regions$geo
}
\keyword{datasets}
