% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edashop-package.R
\docType{data}
\name{price_gradient}
\alias{price_gradient}
\title{Housing Price Gradients in Capital Cities in Italy.}
\format{
A data frame with 112 rows and 21 variables:
\describe{
\item{capital_name}{Provincial Capitals}
\item{grad_existing}{Housing prices gradient (existing units)}
\item{grad_new}{Housing prices gradient (new units)}
\item{population}{Population (no.)}
\item{log_population}{(LOG) Population (No.)}
\item{immigrants}{Immigrants (No.)}
\item{log_immigrants}{(LOG) Immigrants}
\item{emp_retail_tourism}{Employees in retail and tourism (No.)}
\item{emp_service}{Employees in services (No.)}
\item{female_emp_rate}{Female employment rate (\%)}
\item{emp_rate}{Employment rate (\%)}
\item{income_per_capita}{Per capita income (euro)}
\item{log_income_per_capita}{(LOG) Per capita income}
\item{density}{Urban density (inhab/SqKmq)}
\item{log_density}{(LOG) Urban density}
\item{transit_per_capita}{Per capita public transport availability rate}
\item{distance}{Distance between Center and Periphery  (Km)}
\item{housing_surface}{Housing units surface (sqm)}
\item{mean_alt}{Mean altitude (M SL)}
\item{housing_units}{Housing units (No.)}
\item{building_density}{Building density (housing units/res. bld)}
}
}
\source{
Antoniucci, Valentina, and Giuliano Marella. "Housing price gradient and immigrant population: Data from the Italian real estate market." Data in Brief 16 (2018): 794-798. \url{https://doi.org/10.1016/j.dib.2017.12.018}
}
\usage{
data(price_gradient)
}
\description{
These data were collected by Antoniucci and Marella to analyze the correlation between the housing price gradient and the immigrant population
in Italy during 2016. The data may also be useful in other statistical analyses, be they on the real estate market or in other branches of the
social sciences. The data relate to 112 Italian provincial capitals and provide accurate information on urban structure, and specifically on
urban density.
}
\examples{
 data(price_gradient)
 gradient <- price_gradient$grad_existing
}
\keyword{datasets}
