% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edashop-package.R
\docType{package}
\name{edashop-package}
\alias{edashop}
\alias{edashop-package}
\title{edashop: A Package for a Workshop on Exploratory Data Analysis}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A package with templates and data useful to teach a workshop on Exploratory Data Analysis using `R`.
}
\references{
\url{https://github.com/paezha/edashop}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/paezha/edashop}
  \item \url{https://paezha.github.io/edashop/}
  \item Report bugs at \url{https://github.com/paezha/edashop/issues}
}

}
\author{
\strong{Maintainer}: Antonio Paez \email{paezha@mcmaster.ca} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})


Antonio Paez, School of Earth, Environment and Society, McMaster University \email{paezha@mcmaster.ca}
}
\keyword{internal}
